--- onboardings/domain/src/main/java/io/temporal/onboardings/domain/orchestrations/Potato.java
@@ -1,6 +1,0 @@
-package io.temporal.onboardings.domain.orchestrations;
-
-import io.temporal.workflow.WorkflowInterface;
-
-@WorkflowInterface
-public interface Potato {}
